requirements.txt
asgiref==3.8.1
coverage==7.6.11
Django==5.1.6
django-allauth==65.4.1
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.9.0
sqlparse==0.5.3
